backend/requirements.txt
fastapi>=0.116,<1.0
httpx>=0.28,<1.0
openai>=1.109,<2.0
python-dotenv>=1.1,<2.0
uvicorn[standard]>=0.35,<1.0
